+++ app/src/main/java/com/techcourse/config/ManualHandlerAdapter.java
@@ -3,7 +3,7 @@ package com.techcourse.config;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.HandlerAdapter;
-import nextstep.mvc.controller.asis.Controller;
+import nextstep.mvc.controller.simple.Controller;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
