+++ app/src/main/java/com/techcourse/config/ManualHandlerMapping.java
@@ -3,8 +3,8 @@ package com.techcourse.config;
 import com.techcourse.controller.*;
 import jakarta.servlet.http.HttpServletRequest;
 import nextstep.mvc.HandlerMapping;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.controller.asis.ForwardController;
+import nextstep.mvc.controller.simple.Controller;
+import nextstep.mvc.controller.simple.ForwardController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
