+++ mvc/src/main/java/nextstep/mvc/controller/annotation/AnnotationHandlerMapping.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.controller.annotation;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.lang.reflect.Constructor;
