+++ mvc/src/main/java/nextstep/mvc/controller/simple/Controller.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.asis;
+package nextstep.mvc.controller.simple;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
